adding new program (gyroscope) and refactoring

+++ src/altimu-10-v5.c
@@ -3,11 +3,10 @@
 #include "altimu-10-v5.h"
 
 
-
 float pressure_convert(struct pressure press){
     float converted_value;
     
-    converted_value = (float)((press.MSB<<16) + (press.LSB_H<<8) + press.LSB_L)/4096;
+    converted_value = (float)((int32_t)((press.MSB<<16) + (int32_t)(press.LSB_H<<8) + (int32_t)press.LSB_L))/4096;
     
     return converted_value;
 }
